tensorflow/mgaiduk_recall/requirements.txt
certifi
charset-normalizer
decorator
idna
py
requests
retry
urllib3
numpy
pandas
grpc-google-iam-v1
grpcio
grpcio-tools
redis
pyyaml
cloud-tpu-client
tensorflow==2.11.0
tensorflow_io
tensorflow_addons
tensorflow_recommenders